nettoyage de certains problemes + utilisation des methodes de shared
TODO => liste generique dans shared
     => nbreDigit dans shared
     => pb de compil dans p61 a corriger
     => a partir de p62 => a nettoyer (p100, p102, p107, p205, p206 deja
	     fait)

+++ shared/nbre.h
@@ -5,6 +5,10 @@
 #include <stdio.h>
 #include <math.h>
 
-bool isPrime(unsigned int n);
+/*
+ * add nbreDigit
+ * */
+
+bool isPrime(unsigned long n);
 
 #endif
